requirements-core.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
cryptography==44.0.3
cymem==2.0.11
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
fsspec==2025.9.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
ninja==1.13.0
numpy==2.3.3
opencv-python-headless==4.11.0.86
orjson==3.11.3
packaging==25.0
pandas==2.3.2
phonenumbers==9.0.13
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
psutil==7.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydicom==3.0.1
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.4
pytest==8.4.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
rfc3986==2.0.0
rich==14.1.0
ruff==0.12.12
safehttpx==0.1.6
scikit-image==0.25.2
scipy==1.16.1
semantic-version==2.10.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.3
sympy==1.14.0
tifffile==2025.8.28
tldextract==5.3.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.22
psycopg[binary]==3.2.2
